# docs/swagger.yaml
basePath: /
definitions:
  apierrors.ErrorDetail:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  handlers.AuthRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  handlers.AuthResponse:
    properties:
      jwt:
        type: string
      refresh_token:
        type: string
    type: object
  handlers.CartItemResponse:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      price:
        description: Price in cents
        type: integer
      quantity:
        type: integer
      stock:
        type: integer
    type: object
  handlers.CartResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/handlers.CartItemResponse'
        type: array
      user_id:
        type: integer
    type: object
  handlers.CreateCartRequest:
    properties:
      item_id:
        type: integer
      quantity:
        type: integer
    type: object
...
  models.Cart:
    properties:
      item_id:
        type: integer
      quantity:
        type: integer
      user_id:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API for managing items, carts, orders, and authentication.
  title: Checkout API
  version: "1.0"
paths:
...
  /login:
    post:
      description: Logins the user and gives tokens if correncly loginned
      parameters:
      - description: The login details
        in: body
        name: reqBody
        required: true
        schema:
          $ref: '#/definitions/handlers.AuthRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apierrors.ErrorDetail'
      summary: Logins the user
      tags:
      - Auth
...
  /signup:
    post:
      description: Creates(signup) a user for later login use
      parameters:
      - description: The login details
        in: body
        name: reqBody
        required: true
        schema:
          $ref: '#/definitions/handlers.AuthRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apierrors.ErrorDetail'
      summary: Creates a user
      tags:
      - Auth
...
  /user/cart:
    delete:
      description: Removes the user's cart entirely
      parameters:
      - description: The user's ID
        in: header
        name: X-User-ID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apierrors.ErrorDetail'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apierrors.ErrorDetail'
      summary: Removes the user's cart
      tags:
      - Cart
    get:
      description: Gets the given user's cart info
      parameters:
      - description: The user's ID
        in: header
        name: X-User-ID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.CartResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apierrors.ErrorDetail'
      summary: Gets the user's cart
      tags:
      - Cart
  /user/cart/items/{item_id}:
    delete:
      description: Removes the item from the cart entirely
      parameters:
      - description: The user's ID
        in: header
        name: X-User-ID
        required: true
        type: integer
      - description: The item's ID
        in: path
        name: item_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apierrors.ErrorDetail'
      summary: Removes an item from the cart
      tags:
      - Cart
    patch:
      description: Updates the quantity of the given item for the user
      parameters:
      - description: The user's ID
        in: header
        name: X-User-ID
        required: true
        type: integer
      - description: The item's ID
        in: path
        name: item_id
        required: true
        type: integer
      - description: The item's quantity
        in: body
        name: quantity
        required: true
        schema:
          type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apierrors.ErrorDetail'
      summary: Updates the quantity of an item
      tags:
      - Cart
  /user/carts:
    post:
      description: Creates a cart for the user with the given data
      parameters:
      - description: The user's ID
        in: header
        name: X-User-ID
        required: true
        type: integer
      - description: Requested body
        in: body
        name: reqBody
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateCartRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Cart'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apierrors.ErrorDetail'
      summary: Creates a cart for the user
      tags:
      - Cart
...
swagger: "2.0"